+++ backend/lost/db/access.py
@@ -1288,6 +1288,13 @@ class DBMan(object):
             AND a.idx = i.anno_task_id \
             AND i.img_path LIKE '%{search_str}%';"
         return self.session.execute(text(sql))
+
+    def get_search_images_in_annotask_list(self, anno_task_ids, search_str=""):
+        anno_task_list = ",".join([str(x) for x in anno_task_ids])
+        sql = f"SELECT i.idx, i.anno_task_id, i.img_path, a.name FROM anno_task a, image_anno i \
+            WHERE a.idx in ({anno_task_list}) \
+            AND i.img_path LIKE '%{search_str}%';"
+        return self.session.execute(text(sql))
     
     def get_search_images_in_annotask(self, annotask_idx, search_str=""):
         sql = f"SELECT i.idx, i.anno_task_id, i.img_path, a.name FROM anno_task a, image_anno i \