+++ CHANGELOG.md
@@ -82,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Save changed and created annotations instantly to backend
 ### Fixed
 - Data Export Webservices: Fixed Permissions
+- Load img in dataset review: Admins are always allowed to review dataset from now on
 - PipeEngine bug: Created wrong pipe graph, when first element in pe list was 
   not first element in pipeline graph
 - Fixed raw sql in access to be compatible with postgresql